requirements.txt
absl-py==2.1.0
accelerate==0.21.0
ai21==2.8.0
ai21-tokenizer==0.11.2
aider-chat==0.62.1
aiofiles==23.2.1
aiogram==3.8.0
aiohappyeyeballs==2.4.3
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.14.0
altair==5.2.0
annotated-types==0.7.0
anthropic==0.39.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appwrite==7.0.0
APScheduler==3.6.3
archspec @ file:///home/conda/feedstock_root/build_artifacts/archspec_1699370045702/work
asgiref==3.8.1
astunparse==1.6.3
async-timeout==4.0.3
asyncio==3.4.3
attrs==24.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
boltons @ file:///home/conda/feedstock_root/build_artifacts/boltons_1703154663129/work
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1695989934239/work
build==1.2.2.post1
CacheControl==0.14.1
cachetools==4.2.2
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==2.0.12
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.7
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
coloredlogs==15.0.1
conda @ file:///Users/runner/miniforge3/conda-bld/conda_1701731166956/work
conda-libmamba-solver @ file:///home/conda/feedstock_root/build_artifacts/conda-libmamba-solver_1702406360642/work/src
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1691048088238/work
conda_package_streaming @ file:///home/conda/feedstock_root/build_artifacts/conda-package-streaming_1691009212940/work
ConfigArgParse==1.7
contourpy==1.2.0
cryptography==43.0.3
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.20.0
defusedxml==0.7.1
Deprecated==1.2.15
deprecation==2.1.0
diff-match-patch==20241021
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
Django==5.1.6
dnspython==2.7.0
docstring_parser==0.16
durationpy==0.9
ecdsa==0.19.0
einops==0.4.1
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.2.0
Faker==33.1.0
fastapi==0.109.2
ffmpy==0.3.2
filelock==3.16.1
firebase-admin==5.0.0
flake8==7.1.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.49.0
fpdf2==2.7.8
frozenlist==1.5.0
fsspec==2024.10.0
gast==0.6.0
gcloud==0.18.3
gitdb==4.0.11
GitPython==3.1.43
google-api-core==1.34.1
google-api-python-client==2.151.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.14.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
gotrue==1.3.1
gradio==3.41.2
gradio_client==0.5.0
grep-ast==0.3.3
grpcio==1.67.1
grpcio-status==1.48.2
gspread==6.1.2
gTTS==2.5.3
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
h5py==3.12.1
hpack==4.0.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.25.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
importlib_metadata==7.2.1
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.4
jiter==0.7.0
joblib==1.4.2
json5==0.9.25
jsonpatch @ file:///home/conda/feedstock_root/build_artifacts/jsonpatch_1695536281965/work
jsonpointer @ file:///Users/runner/miniforge3/conda-bld/jsonpointer_1695397382064/work
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
keras==3.7.0
kiwisolver==1.4.5
kubernetes==31.0.0
langchain==0.3.12
langchain-chroma==0.1.4
langchain-community==0.3.12
langchain-core==0.3.25
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.3
langsmith==0.1.147
libclang==18.1.1
libmambapy @ file:///Users/runner/miniforge3/conda-bld/mamba-split_1702310516100/work/libmambapy
lightning-utilities==0.10.1
litellm==1.51.2
magic-filter==1.0.12
Mako==1.3.6
mamba @ file:///Users/runner/miniforge3/conda-bld/mamba-split_1702310516100/work/mamba
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.21.3
matplotlib==3.8.3
mccabe==0.7.0
mdurl==0.1.2
menuinst @ file:///Users/runner/miniforge3/conda-bld/menuinst_1702317138933/work
mido==1.3.2
mixpanel==4.10.1
ml-dtypes==0.4.1
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.2.1
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.2.3
onnxruntime==1.16.3
openai==1.12.0
opencv-contrib-python==4.8.0.74
opencv-python==4.10.0.84
openpyxl==3.1.2
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
opt_einsum==3.4.0
optree==0.13.1
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandas-stubs==2.2.3.241009
passlib==1.7.4
pathspec==0.12.1
pdfminer.six==20231228
pdfplumber==0.11.4
peft==0.11.1
pexpect==4.9.0
Pillow==10.1.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1701708255999/work
plotly==5.18.0
pluggy==1.5.0
postgrest==0.11.0
posthog==3.7.0
pretty-midi==0.2.10
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==4.25.6
psutil==6.1.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.1
pycosat @ file:///Users/runner/miniforge3/conda-bld/pycosat_1696355969704/work
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.6.1
pydantic-extra-types==2.10.0
pydantic-settings==2.6.1
pydantic_core==2.16.2
pydeck==0.9.1
pydub==0.25.1
pyflakes==3.2.0
pygame==2.6.0
pygit2==1.12.2
Pygments==2.18.0
PyJWT==2.8.0
pypandoc==1.14
pyparsing==3.1.1
PyPDF2==3.0.1
pypdfium2==4.30.0
pyperclip==1.9.0
PyPika==0.48.9
pyproject_hooks==1.2.0
Pyrebase4==4.8.0
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
pytest==8.3.4
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-jwt==4.1.0
python-multipart==0.0.7
python-slugify==8.0.3
python-telegram-bot==20.6
pytorch-lightning==1.9.4
pytz==2021.1
PyYAML==6.0.2
realtime==1.0.6
referencing==0.35.1
regex==2024.9.11
reportlab==4.3.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.9.3
rpds-py==0.20.1
rsa==4.9
ruamel.yaml @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml_1699007390921/work
ruamel.yaml.clib @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml.clib_1695996997220/work
safetensors==0.4.3
scikit-learn==1.6.0
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==3.3.1
sentencepiece==0.2.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.12.1
soupsieve==2.6
SpeechRecognition==3.11.0
SQLAlchemy==2.0.28
sqlparse==0.5.3
starlette==0.36.3
storage3==0.6.1
streamlit==1.31.0
StrEnum==0.4.15
stripe==7.6.0
supabase==1.2.0
supafunc==0.2.3
sympy==1.12
tenacity==8.4.2
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
text-unidecode==1.3
tf_keras==2.18.0
threadpoolctl==3.5.0
tiktoken==0.8.0
timm==0.9.2
tokenizers==0.19.1
toml==0.10.2
tomli==2.2.1
toolz==0.12.1
torch==2.1.0
torchmetrics==1.3.1
torchsde==0.2.5
torchvision==0.16.0
tornado==6.4.1
tqdm==4.66.6
trampoline==0.1.2
transformers==4.42.3
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
trl==0.9.4
truststore @ file:///home/conda/feedstock_root/build_artifacts/truststore_1694154605758/work
tweepy==4.14.0
typer==0.15.1
types-pytz==2024.2.0.20241003
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tyro==0.8.5
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.27.1
uvloop==0.21.0
validators==0.34.0
watchdog==6.0.0
watchfiles==0.24.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
wrapt==1.17.0
xxhash==3.4.1
yarl==1.17.1
zipp==3.20.2
zstandard==0.22.0
